Added Saved & Watched use cases to builder

+++ src/main/java/view/ToolBarView.java
@@ -110,7 +110,7 @@ public class ToolBarView extends JToolBar {
         setupButton(addToWatchedButton, AppIcon.WATCHED_32.getIcon());
         setupButton(removeFromWatchedButton, AppIcon.WATCHED_REMOVE_32.getIcon());
 
-        Component spacer = Box.createHorizontalStrut((int) Math.round(BUTTON_SIZE * 4));
+        Component spacer = Box.createHorizontalStrut(Math.round(BUTTON_SIZE * 4));
 
         this.add(spacer);
         this.add(searchView);
@@ -122,7 +122,6 @@ public class ToolBarView extends JToolBar {
      * @param icon The icon of the button.
      */
     private void setupButton(JButton button, Icon icon) {
-        // button.setMaximumSize(new Dimension(BUTTON_SIZE, BUTTON_SIZE));
         button.setIcon(icon);
 
         this.add(button);
